moved parsing methods to DaoUtils
